Update test code

+++ benchmark.py
@@ -34,33 +34,43 @@ def run_get(id_):
     return (id_, code, delta.total_seconds())
 
 
+def restart_repo(p_nfd, p_repo):
+    print('Restarting repo')
+    if p_repo:
+        p_repo.send_signal(signal.SIGINT)
+        try:
+            p_repo.wait(1)
+        except TimeoutExpired:
+            p_repo.kill()
+
+    if p_nfd:
+        p_nfd.send_signal(signal.SIGINT)
+        try:
+            p_nfd.wait(1)
+        except TimeoutExpired:
+            p_nfd.kill()
+
+    call(['killall', 'nfd', 'ndn-repo-ng'])
+    sleep(1)
+
+    print('Starting nfd')
+    p_nfd = Popen(['nfd'], stdout=DEVNULL, stderr=DEVNULL)
+    sleep(5)
+    if p_nfd.poll() is not None:
+        print('FAILED to start nfd')
+
+    print('Starting repo')
+    p_repo = Popen(['build/ndn-repo-ng', '-c', 'repo-0.conf'], stdout=DEVNULL)
+    sleep(5)
+    if p_repo.poll() is not None:
+        print('FAILED to start repo')
+
+    return p_nfd, p_repo
+
+
 def run_test(size):
     reset_repo()
     p_nfd, p_repo = None, None
-
-    def restart_repo(p_nfd, p_repo):
-        print('Restarting repo')
-        if p_repo:
-            p_repo.send_signal(signal.SIGINT)
-            try:
-                p_repo.wait(1)
-            except TimeoutExpired:
-                p_repo.kill()
-
-        if p_nfd:
-            p_nfd.send_signal(signal.SIGINT)
-            try:
-                p_nfd.wait(1)
-            except TimeoutExpired:
-                p_nfd.kill()
-
-        p_nfd = Popen(['nfd'], stdout=DEVNULL, stderr=DEVNULL)
-        sleep(5)
-
-        p_repo = Popen(['build/ndn-repo-ng', '-c', 'repo-0.conf'], stdout=DEVNULL)
-        sleep(5)
-
-        return p_nfd, p_repo
 
     with tempfile.NamedTemporaryFile() as tfile:
         print(f'Testing with {size}')
@@ -73,12 +83,17 @@ def run_test(size):
 
         code = call([
             'build/tools/ndnputfile',
+            '-s', '8192',
             '/example/repo',
             '/example/data/1',
             tfile.name
         ])
 
         print(f'return code: {code}')
+
+        if code != 0:
+            print('Put fail. exit test')
+            return []
 
         print(f'Putfile {size} done')
 
@@ -94,8 +109,18 @@ def run_test(size):
                 for values in results
             ]
 
-        p_nfd.kill()
-        p_repo.kill()
+        if p_repo:
+            p_repo.send_signal(signal.SIGINT)
+            try:
+                p_repo.wait(1)
+            except TimeoutExpired:
+                p_repo.kill()
+        if p_nfd:
+            p_nfd.send_signal(signal.SIGINT)
+            try:
+                p_nfd.wait(1)
+            except TimeoutExpired:
+                p_nfd.kill()
 
 sizes = [2 ** i for i in range(10, 30 + 1)]
